Updated min and max GC-Content for long DNA

+++ src/DNAnalyzer/ProteinAnalysis.java
@@ -22,8 +22,10 @@ import java.util.ArrayList;
 public class ProteinAnalysis {
 
   /**
-   * Prints high coverage regions. High coverage regions are regions of a DNA sequence that code for
-   * a protein and have a relatively high proportion of guanine and cytosine nucleotides to the 4
+   * Prints high coverage regions. High coverage regions are regions of a DNA
+   * sequence that code for
+   * a protein and have a relatively high proportion of guanine and cytosine
+   * nucleotides to the 4
    * nucleotide bases (45-60% GC-content).
    *
    * @param geneList
@@ -40,8 +42,8 @@ public class ProteinAnalysis {
     for (final String gene : geneList) {
 
       // High GC content range
-      final double MIN_GC_CONTENT = 0.40;
-      final double MAX_GC_CONTENT = 0.60;
+      final float MIN_GC_CONTENT = 0.40f;
+      final float MAX_GC_CONTENT = 0.60f;
       if ((p.getGCContent(gene) > MIN_GC_CONTENT) && (p.getGCContent(gene) < MAX_GC_CONTENT)) {
         System.out.println(count + ". " + gene);
         count++;
@@ -50,12 +52,14 @@ public class ProteinAnalysis {
   }
 
   /**
-   * Prints the longest protein in the DNA sequence along with its length. Longer genes are most
-   * susceptible to disease implications and are especially linked to neurodevelopmental disorders
+   * Prints the longest protein in the DNA sequence along with its length. Longer
+   * genes are most
+   * susceptible to disease implications and are especially linked to
+   * neurodevelopmental disorders
    * (e.g., autism).
    *
    * @see
-   *     https://www.spectrumnews.org/opinion/viewpoint/length-matters-disease-implications-for-long-genes/
+   *      https://www.spectrumnews.org/opinion/viewpoint/length-matters-disease-implications-for-long-genes/
    * @category Properties
    * @param proteinList The list of proteins in the DNA sequence
    */
